requirements.txt
Flask==2.0.3
matplotlib==3.5.1
numpy==1.21.2
pandas==1.4.1
Pillow==9.0.1
torch==1.10.2
torchvision==0.11.3
gunicorn==20.1.0

